Add functionality to calculate average rate of a book, minor visual changes to book details view

+++ LibrApp2/Models/Review.cs
@@ -15,7 +15,7 @@ namespace LibrApp2.Models
         public string Content { get; set; }
 
         [Range(1, 10, ErrorMessage = "Please enter a value between {1} and {2}")]
-        public int Rate { get; set; }
+        public byte Rate { get; set; }
 
         public int BookId { get; set; }
         public int UserProfileId { get; set; }
